[CHORE]#263 : app 모듈 내 패키지 분리 및 파일 분리, MainActivity 내 MainScreen 분리, MainScreen 내 Component 분리,
