+++ app/src/main/kotlin/com/bff/wespot/main/component/BottomNavigationTab.kt
@@ -1,0 +1,95 @@
+package com.bff.wespot.main.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.bff.wespot.designsystem.theme.StaticTypeScale
+import com.bff.wespot.designsystem.theme.WeSpotThemeManager
+import com.bff.wespot.main.model.BottomBarDestinations
+import com.bff.wespot.ui.util.clickableSingle
+import com.ramcosta.composedestinations.spec.NavGraphSpec
+
+@Composable
+internal fun BottomNavigationTab(
+    selectedNavigation: NavGraphSpec,
+    onNavigationSelected: (NavGraphSpec) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    NavigationBar(
+        containerColor = WeSpotThemeManager.colors.naviColor,
+        modifier = modifier,
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.SpaceAround,
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            BottomBarDestinations.entries.forEach { destination ->
+                TabItem(
+                    icon = painterResource(id = destination.icon),
+                    emptyIcon = painterResource(id = destination.emptyIcon),
+                    title = stringResource(id = destination.title),
+                    description = stringResource(id = destination.title),
+                    selected = selectedNavigation == destination.screen,
+                ) {
+                    onNavigationSelected(destination.screen)
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun RowScope.TabItem(
+    icon: Painter,
+    emptyIcon: Painter,
+    title: String,
+    description: String,
+    selected: Boolean = false,
+    onClick: () -> Unit,
+) {
+    Box(
+        modifier = Modifier
+            .size(80.dp)
+            .weight(1f)
+            .clickableSingle { onClick.invoke() },
+        contentAlignment = Alignment.Center,
+    ) {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Icon(
+                painter = if (selected) {
+                    icon
+                } else {
+                    emptyIcon
+                },
+                contentDescription = description,
+            )
+
+            Text(
+                text = title,
+                style = StaticTypeScale.Default.body9,
+                color = if (selected) {
+                    WeSpotThemeManager.colors.abledIconColor
+                } else {
+                    WeSpotThemeManager.colors.disableIcnColor
+                },
+            )
+        }
+    }
+}
