+++ app/src/main/kotlin/com/bff/wespot/main/model/NavigationBarPosition.kt
@@ -1,4 +1,4 @@
-package com.bff.wespot
+package com.bff.wespot.main.model
 
 internal enum class NavigationBarPosition {
     BOTTOM, TOP,
