+++ app/src/main/kotlin/com/bff/wespot/main/state/MainSideEffect.kt
@@ -1,6 +1,6 @@
-package com.bff.wespot.state
+package com.bff.wespot.main.state
 
-import com.bff.wespot.model.VersionUpdateType
+import com.bff.wespot.main.model.VersionUpdateType
 
 sealed interface MainSideEffect {
     data class ShowVersionUpdateDialog(val versionUpdateType: VersionUpdateType): MainSideEffect
