+++ app/src/main/kotlin/com/bff/wespot/main/state/MainUiState.kt
@@ -1,4 +1,4 @@
-package com.bff.wespot.state
+package com.bff.wespot.main.state
 
 import com.bff.wespot.model.common.Restriction
