+++ app/src/main/kotlin/com/bff/wespot/main/viewmodel/MainViewModel.kt
@@ -1,8 +1,8 @@
-package com.bff.wespot.viewmodel
+package com.bff.wespot.main.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.bff.wespot.MainScreenNavArgs
+import com.bff.wespot.main.MainScreenNavArgs
 import com.bff.wespot.analytic.AnalyticsEvent
 import com.bff.wespot.analytic.AnalyticsHelper
 import com.bff.wespot.common.util.AppVersionUtils.VersionCompareResult
@@ -14,11 +14,11 @@ import com.bff.wespot.domain.repository.user.UserRepository
 import com.bff.wespot.domain.usecase.CacheProfileUseCase
 import com.bff.wespot.domain.util.DataStoreKey
 import com.bff.wespot.domain.util.RemoteConfigKey
-import com.bff.wespot.model.VersionUpdateType
+import com.bff.wespot.main.model.VersionUpdateType
 import com.bff.wespot.model.serverDriven.OnBoardingCategory
-import com.bff.wespot.state.MainAction
-import com.bff.wespot.state.MainSideEffect
-import com.bff.wespot.state.MainUiState
+import com.bff.wespot.main.state.MainAction
+import com.bff.wespot.main.state.MainSideEffect
+import com.bff.wespot.main.state.MainUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.firstOrNull
