+++ app/src/main/kotlin/com/bff/wespot/navigation/navigator/NavigatorImpl.kt
@@ -1,11 +1,13 @@
-package com.bff.wespot
+package com.bff.wespot.navigation.navigator
 
 import android.content.ActivityNotFoundException
 import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
+import com.bff.wespot.BuildConfig
 import com.bff.wespot.auth.AuthActivity
+import com.bff.wespot.main.MainActivity
 import com.bff.wespot.navigation.Navigator
 import com.bff.wespot.navigation.util.buildIntent
 import com.kakao.sdk.common.util.KakaoCustomTabsClient
