+++ app/src/main/kotlin/com/bff/wespot/notification/PushNotificationService.kt
@@ -1,4 +1,4 @@
-package com.bff.wespot
+package com.bff.wespot.notification
 
 import android.app.Notification
 import android.app.NotificationManager
@@ -7,6 +7,7 @@ import android.content.Intent
 import android.media.RingtoneManager
 import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
+import com.bff.wespot.R
 import com.bff.wespot.analytic.AnalyticsEvent
 import com.bff.wespot.analytic.AnalyticsHelper
 import com.bff.wespot.common.CHANNEL_ID
