+++ app/src/main/kotlin/com/bff/wespot/splash/SplashActivity.kt
@@ -8,7 +8,7 @@ import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import com.bff.wespot.PushNotificationService
+import com.bff.wespot.notification.PushNotificationService
 import com.bff.wespot.R
 import com.bff.wespot.navigation.Navigator
 import com.bff.wespot.navigation.util.EXTRA_DATE
